[bot_update] Perform our own cleanup.
Currently, "bot_update" relies on a BuildBot cleanup mechanism and, to a
lesser extent, the standard BuildBot directory layout. Both of these are
problematic when projecting it into other circumstances, notably
"remote_run" and LUCI.

Have "bot_update" handle its own cleanup. It will now choose a cleanup
directory within the hierarchy of its checkout, and explicitly purge it
prior to execution if it exists. This enforces its expected behavior in
all circumstances and removes its expectations of the greater checkout
layout.

BUG=chromium:725631
TEST=expectations
R=hinoka@chromium.org, iannucci@chromium.org

Change-Id: I664434c542a25aaa7ff3eac216208a2425730fde