Fix redist packaging for VS 2017 Update 3 Preview 2
The second preview of VS 2017 Update 3 removed the MFC DLLs from the
redist directory. Luckily it appears that we don't actually need them.

The directories were also renamed from VC150 to VC141. This change
switches to VC* to make it more robust, enabled by the use of glob added
a few changes ago.

crrev.com/2938453003, which tests a package that was created with this
updated script, passed.

BUG=683729
R=scottmg@chromium.org

Change-Id: I4ae5dec949d16a83e1751e3d2ff7bd10a78b680f