ï»¿# Enter a description of the change. Update VC++ packaging script to not package vctip.exe
Some build machines end up running vctip.exe for some reason and then
it doesn't shut down in a timely manner, leaving locks on directories.
This changes the packaging script so that vctip.exe is not packaged, and
therefore won't be run.

vctip.exe is the "MicrosoftÂ® VC compiler and tools experience
improvement data uploader" and it presumably runs automatically as part
of running the compiler, but it's not clear what triggers it, however
omitting it should be safe.

BUG=735226

Change-Id: Ie6af562def6214a5bb130ccc09c732efc1769bcd