gclient: avoid adding potentially duplicate entries to work queue
The main goal is to provide a migration path, where we can add
conditional src-internal entry to src/DEPS, and have it still
work on checkouts where .gclient pulls src-internal, provided
the condition evaluates to False.

The migration path is then to remove the .gclient entry, and
at the same time flip the condition to True (e.g. by overriding
a variable).

Bug: 570091
Change-Id: I9b9850a644463ab0b1f368d65a5cd5f47cf7be97