Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> a

FWIW, I agree that core modules should aim to be as precise as possible, and in this case it seems trivially possible (since the recipe is already started with the change and the patchset numbers).

Rietveld didn't allow for this sort of precision, but Gerrit does (one of the many reasons we're moving to Gerrit). Any new support for Gerrit should support Gerrit in the best way we can.

That said, if this is a blocker for your particular application and you don't want to do the work to implement the solution for the core module right now, you could add some quick code (maybe using the gerrit module directly) to your recipe to unblock it during the prototype phase. Once you're happy with the prototype behavior, you can work on adding proper support to the core module when productionizing the prototype code.

My $0.02

Patch-set: 2
CC: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>