Update patch set 12
Patch Set 12: Commit-Queue+1

PTAL again. Made a few small changes. Now I added a flag in bot_update that controls this. So it will be off by default until downstream recipes switch it on. Like that we prevent non-trivial rolls.

Also, I added a comment about the gerrit query in the "without patch" version. I think it's never necessary as all the revisions in the second run are fixed to the got_revisions of the first run. I.e. we never need to know the branch. But I'll leave that to a separate CL.

Patch-set: 12
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1