Always copy "python.bat".
This change should be safe, since batch files are not generally
write-contended, and "gclient" is run atomically (other things depend on
this).

The mechanism to detect whether we're downgrading is broken by
"bot_update". While future syncs should not encounter this issue,
current syncs still have no way to resolve it, and are left with a
broken "python.bat" setup.

We remedy this by always installing "python.bat", even if it's already
there. This is safe, since the intended state of HEAD is for this file
to be present at all times.

BUG=chromium:741650
TEST=None

Change-Id: Ief28217bf89d201d830c46b31e0b4040c51cb322