Remove presubmit check for DISALLOW_* macros
Before C++11, we were using a hack to disable copy constructors or copy
assignment. This hack required the respective macros to be placed in
the private: declarations of a class.
Since we changed these macros to use the C++11 "= delete" syntax
(https://codereview.chromium.org/1123723003/), this requirement is gone.
The comments on the macros are already updated in v8 and chromium:
https://chromium-review.googlesource.com/c/577687
https://chromium-review.googlesource.com/c/578027

R=jochen@chromium.org

Change-Id: I224005a17c3de2569b64622e6cf8c9688665637d