Update patch set 2
Patch Set 2:

(3 comments)

> Patch Set 1: Code-Review+1
> 
> (3 comments)
> 
> Seems fine. What's the expected use case here?
In downstream project we have some cases when .gclient file already generated and we need to provide path to the cache.

In past we used to do "import os" in the .gclient file and pass value from the environment variable. But it's weird and conflicts with syntax validation. With this change we can pass path to cache to gclient without any hacks in the config.

Patch-set: 2
Reviewer: Gerrit User 1128614 <1128614@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>