Update patch set 5
Patch Set 5:

> Patch Set 4: Commit-Queue+2
> 
> (1 comment)
> 
> Thanks for the reviews!  How long should I wait after submitting this before submitting the Chromium CL (tweaking the presubmit hooks) that depends on this?  In other words, do I need to do anything to ensure that this change rolls out to Chromium developers?  Is there a regular schedule on which depot_tools is updated?

depot_tools is auto-updating (sort of, life is actually more complicated) and so you should be able to assume it's effective immediately.

However, there is actually a copy of depot_tools that is pulled in via DEPS, and if you could roll that as part of your src-side change (or before), that would be a good idea.

Patch-set: 5