Update patch set 4
Patch Set 4:

Looks like kjellander@ commented on the bug that he didn't think we should fix this as it's too risky. 

Since I don't know that anyone outside of WebRTC really needs this, I'm inclined to defer to him, but the change looks reasonable to me, so if he was on the fence, I'd be inclined to land it.

Patch-set: 4
Reviewer: Gerrit User 1002075 <1002075@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>