Make gclient.sync() 'with_branch_heads' arg actually do something.
This arg used to be used in a 'sync' call that was removed in
crrev.com/7090f652d180eeff2663e3e3091d5db14123284c, but it was never
applied to the remaining 'sync' call. This enables the flag again, and
changes the default to reflect the fact that it has always behaved as if
set to "True".

R=agable@chromium.org, phajdan.jr@chromium.org

Change-Id: Ice5a592d95a0b09dafddab24769f5edd8835a1c9