Update patch set 2
Patch Set 2: Commit-Queue+1

(1 comment)

It's not quite irrelevant yet. And as it becomes more irrelevant, this is still going to be the one part that has to persist: given a change to a particular repo, which directory should be checked out to that change? Maybe this will eventually be handled by gclient itself (something like --rev=https://some.url/repo@refs/changes/foo, and then automatically subbing that in when it finds a path cloning from that url), but for now the recipe still needs to tell gclient which dep should be customized.

Patch-set: 2
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1