Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Under what circumstances would we want this? Whatever a bot is operating on should be something that a human can replicate, i.e. it should be something with a URL. I'd prefer not to support "purely local" arrangements.

I want to use this for the buildspec recipes, which have different needs than normal "builder" bots. And even in this case, it is working with a real repo, and is completely replicable by humans. In fact, it's the configuration I've been using for manual flattening, which I now am trying to migrate to the bots.

Patch-set: 1
Reviewer: Gerrit User 1002222 <1002222@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>