Bootstrap windows tools even when update is disabled.
We have a problem with bootstraping depot_tools on our bots in downstream
project. We disabled autoupdate of depot_tools because we want to control
it's revision manually. But when updates are disabled, we must call
cipd_bin_setup and bootstrap/win/win_tools.bat explicitly to ensure that
cipd, git, vpython and gsutil are downloaded.

This change makes possible to:

1. Checkout depot_tools on any revision.
2. Disable autoupdates with update_depot_tools_toggle.py.
3. Use any tools from depot_tools (including gclient runhooks).

Bug:
Change-Id: Ic15a25adfd2d61212872c0b8e729023c23849cd8