Update patch set 3
Patch Set 3:

lgtm, this makes sense and looks good.

It is surprising to me that this actually makes a difference, since the re module is supposed to keep a cache of previously-compiled regular expressions on its own anyway. But I guess the list of regexes for watchlists is so big and doesn't obey good cache-access patterns that they get evicted.

Patch-set: 3
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>