Update patch set 1
Patch Set 1:

I was chatting to Robbie and we came up with a much better way to do this. Exposing os.remove() via the InputApi is... undesired. The better thing to do would be to give the InputApi a NamedTemporaryFile-style method. So then the caller could do something like:

    with input_api.NamedTemporaryFile() as f:
      do_things_with_a_file(f)

and the implementation on the InputApi-side would, rather than deleting the file immediately like NamedTemporaryFile(delete=True) does, would use delete=False and store the name of the file in a list of all files created. Then right at the end of the whole presubmit run, it would os.remove all files created this way.

Does that makes sense?

Patch-set: 1
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>