Update patch set 5
Patch Set 5:

How about the approach in patch v5? It's essentially what we were discussing, but it also throws a TypeError if delete=False is passed, as accepting an argument but only when it has a certain value can be confusing as well.

Patch-set: 5
Reviewer: Gerrit User 1111977 <1111977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>