Update patch set 1
Patch Set 1:

I went ahead and made other CLs that correspond to this change as well (all the presubmit scripts that was added in the past). I'll put those out for review as well so that once this is landed, all the others can be landed in quick succession.

Patch-set: 1
Reviewer: Gerrit User 1246322 <1246322@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false