Add more tests for fix_encoding.py
In practice, there's something magical that happens when a python call another
process, even if there's a bash in between, that causse the child process to not
be affect by the lack of LANG variable. This is disconcerting, as if I do:
  export -n LANG
  ./tests/fix_encoding_test.py

FixEncodingTest.test_unicode fails with:
  Traceback (most recent call last):
    File "./tests/fix_encoding_test.py", line 45, in test_unicode
        print self.text
        UnicodeEncodeError: 'ascii' codec can't encode character u'\xe9' in
        position 1: ordinal not in range(128)

Yet all the tests print out \xff without problem. The only difference is if
stdout/stderr is piped. For example, the following pass:
  export -n LANG
  ./tests/fix_encoding_test.py &> a.txt
  cat a.txt

I don't see a way to get around that. The unit tests are still valuable so
committing them.

Test only change.

R=vadimsh@chromium.org
Bug: 787531
Change-Id: I0c97ba7b8625b06017135f48363c55dd4fc40557