depot_tooks: Invoke cros via vpython and add oauth2client to spec.
Problem:
depot_tools/cros (that wraps gclient in cros builds) is failing with
import errors on some bots:
https://luci-milo.appspot.com/p/chromium/builders/luci.chromium.findit/findit_variable/94

This will ensure that a version of oauth2client is availabe when
the script is executed. Hopefully vpython is available across any cros
builders.

And hopefully vpython's env is maintained across execv's:
https://chromium.googlesource.com/chromium/tools/depot_tools/+/master/support/chromite_wrapper#71

Bug: 792733
Change-Id: Id4cfb631f64f96446db7072dd727fa0a32607067