Update patch set 6
Patch Set 6:

This is a bad solution.

You are invoking non-vpython code that has complex dependencies with vpython. The dependencies are managed, but the mechanism to do so is somehow broken. The script in question is attempting to import from chromite/third_party/oath2client/, and is expecting that version, and to track that version in the future.

This is the same issue I'm seeing in attempts to do ChromeOS builds on swarming, and which I've (so far) been unable to hack around.

https://bugs.chromium.org/p/chromium/issues/detail?id=783517

Patch-set: 6