Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Is there any ETA for testing/merging this CL?
> 
> I'm confused. What's your reason for wanting this to land ASAP?
> 
> Note that this CL changes **default** version. Whoever uses this script can override it trivially and pass whatever version they need.

Currently we turning down 40 builders (25% of our fleet) in FRA DC due to b/70563780 caused by gsutil v4.13 use of SSLv3
We are working on workarounds to pin latest version in our scripts. This have to be done in multiple scripts and merged to multiple branches, also we are not owning all of those scripts and global update of gsutil in depot_tools will resolve an issue completely.

Please consider that current version pin 4.13 is fundamentally broken due to use outdated httplib2 and SSLv3, so will fail in each backend where SSLv3 is disabled.

Patch-set: 1