Update patch set 1
Patch Set 1:

Hi Dmytro, how quickly do you need this done? hinoka is out for two weeks so if you need it sooner we'll need to find someone else to help roll this out. This is kind of a scary change to deploy so it will take a lot of care and testing (e.g. following tandrii's and iannucci's suggestion to use the led tool). Anything else you can do on your end to not depend on the default version of this will be a faster and safer option.

Patch-set: 1