Fix broken git rev updating with deps_os entries.
The first lss entry updates correctly (for os=android), but when
the second for os=unix runs, the old revision has already been
overwritten. start_idx, therefore, becomes -1 and update_string
writes in the wrong place. This makes update_string a no-op if
it doesn't find what it's supposed to update.

Bug:chromium:798336
Change-Id: I91bbcb8c437bd6583dd89dd0b643c366835b5ef8